+++ src/Traits/HasMorphOwner.php
@@ -9,22 +9,22 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Ownership\Traits;
+namespace Cog\Laravel\Ownership\Traits;
 
+use Cog\Laravel\Ownership\Observers\OwnableObserver;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Database\Eloquent\Builder;
-use Cog\Ownership\Observers\ModelObserver;
-use Cog\Ownership\Contracts\CanBeOwner as CanBeOwnerContract;
+use Cog\Contracts\Laravel\Ownership\CanBeOwner as CanBeOwnerContract;
 
 /**
  * Class HasMorphOwner.
  *
- * @package Cog\Ownership\Traits
+ * @package Cog\Laravel\Ownership\Traits
  */
 trait HasMorphOwner
 {
     /**
-     * @var \Cog\Ownership\Contracts\CanBeOwner|null
+     * @var \Cog\Contracts\Laravel\Ownership\CanBeOwner|null
      */
     private $defaultOwner;
 
@@ -35,7 +35,7 @@ trait HasMorphOwner
      */
     public static function bootHasMorphOwner()
     {
-        static::observe(new ModelObserver);
+        static::observe(OwnableObserver::class);
     }
 
     /**
@@ -61,7 +61,7 @@ trait HasMorphOwner
     /**
      * Get the model owner.
      *
-     * @return \Cog\Ownership\Contracts\CanBeOwner
+     * @return \Cog\Contracts\Laravel\Ownership\CanBeOwner
      */
     public function getOwner()
     {
@@ -71,7 +71,7 @@ trait HasMorphOwner
     /**
      * Get default owner.
      *
-     * @return \Cog\Ownership\Contracts\CanBeOwner|null
+     * @return \Cog\Contracts\Laravel\Ownership\CanBeOwner|null
      */
     public function defaultOwner()
     {
@@ -81,7 +81,7 @@ trait HasMorphOwner
     /**
      * Set owner as default for entity.
      *
-     * @param \Cog\Ownership\Contracts\CanBeOwner|null $owner
+     * @param \Cog\Contracts\Laravel\Ownership\CanBeOwner|null $owner
      * @return $this
      */
     public function withDefaultOwner(CanBeOwnerContract $owner = null)
@@ -122,7 +122,7 @@ trait HasMorphOwner
     /**
      * Resolve entity default owner.
      *
-     * @return \Cog\Ownership\Contracts\CanBeOwner|null
+     * @return \Cog\Contracts\Laravel\Ownership\CanBeOwner|null
      */
     public function resolveDefaultOwner()
     {
@@ -132,7 +132,7 @@ trait HasMorphOwner
     /**
      * Changes owner of the model.
      *
-     * @param \Cog\Ownership\Contracts\CanBeOwner $owner
+     * @param \Cog\Contracts\Laravel\Ownership\CanBeOwner $owner
      * @return $this
      */
     public function changeOwnerTo(CanBeOwnerContract $owner)
@@ -166,7 +166,7 @@ trait HasMorphOwner
     /**
      * Checks if model owned by given owner.
      *
-     * @param \Cog\Ownership\Contracts\CanBeOwner $owner
+     * @param \Cog\Contracts\Laravel\Ownership\CanBeOwner $owner
      * @return bool
      */
     public function isOwnedBy(CanBeOwnerContract $owner)
@@ -181,7 +181,7 @@ trait HasMorphOwner
     /**
      * Checks if model not owned by given owner.
      *
-     * @param \Cog\Ownership\Contracts\CanBeOwner $owner
+     * @param \Cog\Contracts\Laravel\Ownership\CanBeOwner $owner
      * @return bool
      */
     public function isNotOwnedBy(CanBeOwnerContract $owner)
@@ -193,7 +193,7 @@ trait HasMorphOwner
      * Scope a query to only include models by owner.
      *
      * @param \Illuminate\Database\Eloquent\Builder $query
-     * @param \Cog\Ownership\Contracts\CanBeOwner $owner
+     * @param \Cog\Contracts\Laravel\Ownership\CanBeOwner $owner
      * @return \Illuminate\Database\Eloquent\Builder
      */
     public function scopeWhereOwnedBy(Builder $query, CanBeOwnerContract $owner)
@@ -208,7 +208,7 @@ trait HasMorphOwner
      * Scope a query to only include models by owner.
      *
      * @param \Illuminate\Database\Eloquent\Builder $query
-     * @param \Cog\Ownership\Contracts\CanBeOwner $owner
+     * @param \Cog\Contracts\Laravel\Ownership\CanBeOwner $owner
      * @return \Illuminate\Database\Eloquent\Builder
      */
     public function scopeWhereNotOwnedBy(Builder $query, CanBeOwnerContract $owner)